Monitor the filters on the query page to convert to subcomponents
The previous implementation relied on the change event for the
'add filter' box to be called for componentselect.js last, as the first
events would add the DOM. In order to achieve that this event handler
was always called last, we manipulated the HTML to put the
componentselect.js script last in the head section, instead of a
different order when we use trac's regular `add_script` API.

With the demise of Genshi, this is no longer feasible. Instead we use
the MutationObserver to monitor changes, and convert the component
select boxes after the fact.